OS X compatibility per review comments
