+++ shared/util_unix.go
@@ -1,0 +1,15 @@
+// +build !windows
+
+package shared
+
+import (
+	"os"
+	"syscall"
+)
+
+func GetOwner(fInfo os.FileInfo) (os.FileMode, int, int) {
+	mode := fInfo.Mode()
+	uid := int(fInfo.Sys().(*syscall.Stat_t).Uid)
+	gid := int(fInfo.Sys().(*syscall.Stat_t).Gid)
+	return mode, uid, gid
+}
